refactor: reorganize folder structure
