+++ src/data/static/currencies.json
